MDEV-34348: Review 3 fixups
pushing for buildbot. will reorgainze commits later.

+++ sql/sql_select.cc
@@ -25078,7 +25078,7 @@ static int remove_dup_with_hash_index(THD *thd, TABLE *table,
     (*field_length++)= (*ptr)->sort_length();
 
   if (my_hash_init(key_memory_hash_index_key_buffer, &hash, &my_charset_bin,
-                    file->stats.records, 0, key_length, 0, 0, 0))
+                   (uint) file->stats.records, 0, key_length, 0, 0, 0))
   {
     my_free(key_buffer);
     DBUG_RETURN(1);
